backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.35.0
google-generativeai==0.8.6
chromadb==0.5.3
pypdf==4.2.0
pydantic==2.7.4
pydantic-settings==2.3.4
websockets==12.0
httpx==0.27.0
tiktoken==0.7.0

sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
